requirements.txt
alembic
crewai
crewai-tools
fastapi
langchain-openai
python-dotenv
mysql-connector-python
pydantic-settings
chromadb
duckduckgo-search
playwright
streamlit
pandas
plotly
graphviz
pyvis
markdown
requests
beautifulsoup4
uvicorn
